Remove Adapter._prepare_training_data
Summary:
This method checks that the features of `Observation` objects that are constructed by `observations_from_data` are unique. However, due to the reliance on (approximate) Ax equality checks in `set(observation_features)`, this can be extremely slow and inaccurate. It leads to errors by flagging distinct observation features as duplicates.

Removing it will avoid the faulty errors and should improve runtime, particularly when MapData is utilized.

Differential Revision: D71984539