remove SQLAlchemy from critical import path (#3573)
Summary:
Pull Request resolved: https://github.com/facebook/Ax/pull/3573

Changes to `__init__.py` in FOO caused SQLAlchemy to make its way onto the critical path for `import ax`. Because we want to maintain SQLAlchemy as an optional dependency (for now, this may be rethought in the future) we need to make sure this path can gracefully fail.

See breakage here: https://github.com/facebook/Ax/actions/runs/14099230949/job/39492253946

Reviewed By: esantorella

Differential Revision: D71978009

fbshipit-source-id: 137a83692c6d39685ee57c38307cd52e8d9c6d74