Update Adapter.set_status_quo to support MapData
Summary:
Previously, we would set `Adapter.status_quo` to `None` if the status quo arm had multiple corresponding `Observation` objects at the target trial. This is pretty much always the case if there are `MapMetrics` that are observed at multiple progressions.

This diff updates the logic to check for map keys in the optimization config, and set `Adapter.status_quo` to the `Observation` with the maximal map key value (if there is a single map key).

Reviewed By: Balandat

Differential Revision: D72273393

fbshipit-source-id: df6bbba77bd12d03c6b9980da912fff75e68f221