Make SensitivityAnalysisPlot take in one metric name
Summary:
Erogonomics change. Everyone passes in a single element list where `metric_names=[metric_name]` so we might as well just accept the one.

This was originally written to accept an iterable because ax_parameter_sens takes a list, but there is no computational speedup to batching here so there really is no point in continuing that pattern in SensitivityAnalysisPlot

Differential Revision: D72335119