Disallow None in parameter values, drop None observations in Cast (#3604)
Summary:
Pull Request resolved: https://github.com/facebook/Ax/pull/3604

`None`s in parameters are often used to represent unknown or default values of the status quo parameters. Besides this use case, `None` is not an acceptable value for any of the Ax `ParameterTypes`. If `None`s sneak into the `ObservationFeatures` lower in the stack, they lead to errors.

`Adapter` currently relies on the in-desing point filtering / search space membership checks to remove observations with `None`s.  In this diff, we're changing things slightly, to allow for `Cast` to remove `None` observations as well. This will allow us to simply transform the observation features and ensure that they are `None`-free.

In addition, `Parameter.cast` is updated to raise an error rather than returning back `None` values. This will prevent the users from mistakenly constructing parameters with `None`s in the values.

Reviewed By: sdaulton

Differential Revision: D72400250

fbshipit-source-id: 560d59f3ffe4955db734988c8e5793f4cdb48261