When running a benchmark with a simulated backend, replace real time with simulated time on Trials (#3599)
Summary:

The effects are best illustrated by looking at the x-axes on these plots:

Currently, the plot produced by `get_standard_plots` looks distorted because it uses wall time, and epochs can progress nearly instantaneously and there can be a pause while waiting for a trial to start:

{F1976732706}

With these changes, time is in simulated seconds since the start of time (note the 1969 date).

{F1976732734}

A note about cumulative epochs: If we want to know the best point seen in terms of the best seen as a function of cumulative epochs rather than time, we can continue to use `get_benchmark_result_with_cumulative_steps`. We don't have a utility for plotting individual curves in terms of cumulative epochs or of combining these plots, but we could add that by working with the code in ax/service/utils/report_utils.py (where `get_standard_plots`) is defined.

Another note: I hope to break up `benchmark_replication` to make it less of a monster and

Reviewed By: Balandat

Differential Revision: D72351793