Bug fix: Use simulated backend even when max_parallelism=1 when there is an ESS or variable runtimes (#3600)
Summary:
Pull Request resolved: https://github.com/facebook/Ax/pull/3600

Context: Currently, at odds with a docsting, we use a simulated backend only when max_parallelism > 1, since that is the only time when we need a simulated backend to ensure that events happen in the correct order. However, I don't think early stopping would work without a simulated backend, regardless of parallelism, since `BenchmarkMapMetric` relies on the backend simulator to know what data from not-run steps to discard. Lack of a simulated backend also inhibits proper tracing of cost.

This PR:
* Adds an argument `force_use_simulated_backend` to `BenchmarkRunner`
* Makes it True when constructing a runner and there is an early stopping strategy

Reviewed By: Balandat

Differential Revision: D72391347

fbshipit-source-id: 4ab23f4aa8e2103f3ebfa1d2becda46a99deaf8f