Update heterogeneous noise test
Summary: Separating the Ax and BoTorch changes from D72406752. That this test actually worked is a good example for why we need more validation in the BoTorch test functions. We are basically passing in a 1D input to Hartmann6 that is also outside the `bounds`, but things still run because of broadcasting and lack of validation on the BoTorch side.

Differential Revision: D72513532