Update heterogeneous noise test (#3607)
Summary:
Pull Request resolved: https://github.com/facebook/Ax/pull/3607

Separating the Ax and BoTorch changes from D72406752. That this test actually worked is a good example for why we need more validation in the BoTorch test functions. We are basically passing in a 1D input to Hartmann6 that is also outside the `bounds`, but things still run because of broadcasting and lack of validation on the BoTorch side.

Reviewed By: esantorella

Differential Revision: D72513532

fbshipit-source-id: e28693aa1769042ce57d5bd411b8a6d75f82612c