Fix Cast.transform_observations
Summary:
In D72400250, we started dropping observation features with `None`s in `Cast.transform_observation_features`. This doesn't play well with base `Transform.transform_observations` implementation, since the observation data is unmodified and re-construction of the `Observation` objects fails due to lists of unequal lengths.
This diff fixes the issue by implementing `Cast.transform_observations` that processes each observation in a loop, dropping it completely if the features are dropped. The looping here is ok (will probably go away with the larger refactor) since the underlying transform logic also loops over the objects to process them.

Reviewed By: Balandat

Differential Revision: D72568516