Extract generated points from experiment in RandomAdapter rather than tracking them on RandomGenerator (#3605)
Summary:
Pull Request resolved: https://github.com/facebook/Ax/pull/3605

`RandomGenerator` was keeping track of generated points, and using them for deduplication regardless of the status of the trials generated using these points. In experiments with discrete spaces and many FAILED trials, this eventually leads to all of the search space being included in `generated_points` and the random node not being able to generate any new candidates.

This diff replaces the logic to instead extract the points to deduplicate against in `RandomAdapter`, using the same utility that controls deduplication in `GeneratioNode` level. The reason for handling deduplication within `RandomGenerator` rather than deferring to `GenerationNode` is that the generators operate on the continuous space and they may generate many duplicates (quite cheaply) before finding a point that is not a duplicate after transforming back to the original space. If we were to rely on `GenerationNode`, we would only be allowed 5 attempts to generate a non-duplicate trial, which may result in quickly falling back to Sobol node to generate another possibly duplicate arm.

Reviewed By: lena-kashtelyan

Differential Revision: D72331178

fbshipit-source-id: 37d77ced6a633bb72b4426074c484fff5932bcc7