Add p_feasible column to prepare_arm_data (#3618)
Summary:
Pull Request resolved: https://github.com/facebook/Ax/pull/3618

Adds a new column to the result of prepare_arm_data to be consumed downstream in Analyses like ArmEffects and Scatter. Dispatches to botorch's compute_log_prob_feas_from_bounds to compute.

If an arm is missing a mean for any metric with an outcome constraint its p_feasible row will be NaN.
If an arm is missing a sem 0 will be used instead (ie we will return 1 if the mean is within the bound and 0 otherwise).

In plots which utilize this column we will set some consistent threshold (RFC: sdaulton suggested 0.95) and indicate to the user that this point should be investigated for constraint violation

Reviewed By: Balandat

Differential Revision: D72584699