New unified ArmEffects plot (#3617)
Summary:
Pull Request resolved: https://github.com/facebook/Ax/pull/3617

Plot a point for each arm on the experiment for a single metric.

If this performs well we will reap InSampleEffects and PredictedEffects as this performs the same function, and we will replace the whole arm_effects module with just this unified.py plot (renamed to arm_effects).

Implements all features in the above plots *except* an indication for constraint violation (will implement in a followup to keep discussion limited to that diff).

TODO: Take care of the following in follow up diffs
* Add online/offline tests once D72271218 lands
* Default to Generators.EMPIRICAL_BAYES_THOMPSON if no suitable model is found and use_model_predictions=True
* Add adhoc method `arm_effects_adhoc(experiment: Experiment, metrics: Sequence[str] | None, use_model_predictions: bool, adapter: Adapter | None, trial_index: int | None, additional_arms: Sequence[Arms] | None, labesl: Mapping[str, str] | None) -> list[PlotlyAnalysisCard]`
* Add "halo" for points which may violate constraints
* Add marker for trial status (specifically CANDIDATE trials)
* Add `relativize` flag to init and to adhoc method
* Generate this in UI
* Reap InSampleEffects and PredictedEffects

Reviewed By: sdaulton

Differential Revision: D72422273

fbshipit-source-id: 41929a93929b5701ef0fc4a1a7baeecaf9aa5276