Simplify Adapter.predict implementation (#3622)
Summary:
Pull Request resolved: https://github.com/facebook/Ax/pull/3622

This removes all complicated error handling from the predict method, including the fallback to looking for matching observations in the training data if the predictions fail. With this change, predict just predicts.

I don't know if there was any place in the codebase that relied on these fallbacks. Tests didn't find any.

In the past, `Adapter.gen` would error out if it got an error from `Adapter.predict` (which it always calls). This has recently been updated to catch and discard such errors, to prevent a simple side effect from breaking the main functionality. So, unless there is somewhere in the code where we call predict explicitly with a model that is not capable of it, this shouldn't cause any issues.

Reviewed By: Balandat

Differential Revision: D72676325

fbshipit-source-id: 9984eb9c5bfe51b022c9aee88dded6a6aa7ae29c