Remove full `ObjectiveThreshold` objects from `GeneratorRun.gen_metadata`
Summary:
This diff avoids putting `ObjectiveThreshold` objects into `GeneratorRun.gen_metadata`. T206493044 explains why this is a bad thing to do.

In a previous iteration of this diff, I saved this information in the form of python primitives, but after discussion with sdaulton I stopped saving them altogether as they're not actually used anywhere (their previous use in `get_pareto_optimal_parameters` was bugged and so never had any effect).

D72802998 modifies the storage layer to avoid loading any objective thresholds that were saved in the past.

For posteriority, here some thoughts on the design of a previous iteration:

~~I opted to add a `to_dict()` method to `ObjectiveThreshold`, which stores the key properties. Since we can't store full `Metric` objects this simply stores the metric name instead. If we want to re-construct the full object, we have to grab the `Metric` objects from somewhere else, e.g. the `Experiment` - this is done e.g. in the new `_get_objective_thresholds_override` function.~~

Reviewed By: Cesar-Cardoso

Differential Revision: D72733646

fbshipit-source-id: 8068e10af770653774aacc52c7de48c77aaf4065