Fix Adapter._set_status_quo when there are no map keys (#3635)
Summary:
Pull Request resolved: https://github.com/facebook/Ax/pull/3635

The previous implementation could try to pop from an empty set if the optimization config didn't have any map keys but the tracking metrics did. To stay on the safe side, this diff updates the logic to skip setting status quo if there are no map keys in the optimization config when there are multiple status quo observations.

Reviewed By: bernardbeckerman

Differential Revision: D72869084

fbshipit-source-id: 316af474b482b9dbc212ed0964478b9faa1327b3