Improve relativize logic
Summary:
Slight modification of the relativization logic. _relativize_df takes in the same df for the first arg, then a diff that holds one row per trial with the data to use as control for that trial. Controls are picked with the following preferences:
1. If a trial has its own status quo use that
i. If that status quo has NaN's in its metric value fall back to raw observations
2. If there is only one status quo on the experiment use that
i. If that status quo has NaN's in its metric value fall back to raw observations
3. Raise an Exception

Reviewed By: eonofrey

Differential Revision: D72876921

fbshipit-source-id: ad3b1c02ee589bb3a462b95729f5f9f7388eb2df