Docusaurus: stable docusaurus links to stable ReadTheDocs (#3643)
These were erroneously pointing to the latest version in ReadTheDocs. We changed the API docs in RTD (reflected on version "latest") which broke these links from the stable docs in Docusaurus.

When making retroactive changes to the branch `docusaurus-versions` we must ensure that it won't create merge conflicts with the `main` branch. In this case that means only updating the links in `website/versioned_docs`. This branch still has erroneous links in the top level `docs/` directory but those get overwritten by the latest changes in `main` when deploying the website.