Replace InsampleEffects, PredictedEffects with ArmEffects in PTS
Summary:
As titled. Also add relativize=True to ScatterPlots If these are successful we can reap the InSampleEffects and PredictedEffectsPlot, which are less fully featured versions of this.

Misc. fixes needed for tests to pass, including fallback to EB when use_model_predictions=True but no suitable model found (this is the existing behavior of InSampleEffects)

Reviewed By: eonofrey

Differential Revision: D72745006

fbshipit-source-id: 8eab40cf5e20472c7b7009ab695a5e7eca7957e6