Misc fixes in cv plot (#3652)
Summary:
Pull Request resolved: https://github.com/facebook/Ax/pull/3652

Various misc changes.
## Compute for all metrics if None is provided

This is in line with other plots. Also change refined_metric_name to labels dict so that labels can be applied to all cv plots generated correctly.

## Fix compute_cross_validation_adhoc ignoring passed in Data

Previously this was used just to extract metric names -- presumably we want to actually fit the model with this data

## Make args to compute_cross_validation_adhoc match CrossValidationPlot

As titled. These should not be out of sync

## Make colors match other PlotlyAnalysis
Change the blue here to be the same blue as the other plots, and use the same alpha for the CI

Differential Revision: D72802304