Use AutoTC instead of MinTrials in CenterGN (#3666)
Summary:
Pull Request resolved: https://github.com/facebook/Ax/pull/3666

The `AutoTC` always transitions after generating one GR, whereas `MinTrials` requires the trials to be attached to the experiment before it can transition, which causes issues if we try to generate multiple trials at once before attaching them to the experiment (we either get only one trial or generate multiple center points). `AutoTC` seems like the better option in this case, though it is not perfect either. If we generate and forget to attach that trial to the experiment, we will not generate from Center again. I think this is a fair compromise to unblock the use of CenterGN with Ax API, which supports requesting multiple trials at once.

Reviewed By: lena-kashtelyan

Differential Revision: D73276555

fbshipit-source-id: de51842f83bf36d24acc07c2e7b022ece8fafe69