Use GeneratorRun.best_arm_predictions for best points used for inference value
Summary:
Context:

* Benchmarking has its own best-point logic that is different from Ax's best-point logic, which is essentially writing `best_arm_predictions` to a GeneratorRun
* This discrepancy makes developing and debugging best-point functionality hard.
* Also, current best-point logic is intertwined with Scheduler, which makes it hard to move off of Scheduler

With this change, benchmarking will not have its own best-point logic, but will instead use the values on `GeneratorRun.best_arm_predictions`, for better or for worse. This may mostly make the best-point recommendations worse. That is bad for using the inference trace to benchmark noisy problems, but it is good for developing better best-point functionality.

This PR:

Uses GeneratorRun.best_arm_predictions for best points used for inference value.
* Deletes `BenchmarkMethod.get_best_parameters`
* Consolidates `_get_inference_trace_from_params` and `_get_oracle_trace_from_arms` into one function

Differential Revision: D73148478