Update significant figures of Improvement Report
Summary:
Issues: Improvement report may not helpful if the changes between baseline value and best arm value are small (as we round to 2 digits by default). For example, we see
"Metric improved 10.10% from 0.04 to 0.04" when the actual metrics are 0.0439037704 and 0.0394705458.

This change adaptively find the number of significant figures to display in a comparison message.

If the absolute difference between the two values is >= 1, default to round to 2 digits. 
If the difference between the two values is < 1, we find the first differing digit (first non-zero element of the digits of the abs difference) and round to that digit (up to the 10th digit).

Differential Revision: D73525724