Update significant figures of Improvement Report (#3682)
Summary:
Pull Request resolved: https://github.com/facebook/Ax/pull/3682

Issues: Improvement report may not helpful if the changes between baseline value and best arm value are small (as we round to 2 digits by default). For example, we see
"Metric improved 10.10% from 0.04 to 0.04" when the actual metrics are 0.0439037704 and 0.0394705458.

This change adaptively find the number of significant figures to display in a comparison message.

If the absolute difference between the two values is >= 1, default to round to 2 digits.
If the difference between the two values is < 1, we find the first differing digit (first non-zero element of the digits of the abs difference) and round to that digit (up to the 10th digit).

Reviewed By: dme65

Differential Revision: D73525724

fbshipit-source-id: 38b0ca6e39cc708dd1c8962dc70e2ba5e8eedc85