Make images in docs smaller
Summary:
These were defaulting to filling up the full width of the page, which in some cases made them tall enough to basically be in the only thing visible on the page. Let's make them a bit smaller.

It's a bit annoying that this isn't possible just in markdown, so we change these docs to mdx and use html for sizing.

Differential Revision: D73619790