Make images in docs smaller (#3686)
Summary:
Pull Request resolved: https://github.com/facebook/Ax/pull/3686

These were defaulting to filling up the full width of the page, which in some cases made them tall enough to basically be in the only thing visible on the page. Let's make them a bit smaller.

It's a bit annoying that this isn't possible just in markdown, so we change these docs to mdx and use html for sizing.

Reviewed By: mpolson64

Differential Revision: D73619790

fbshipit-source-id: f49d61fb698ded5c83e5a14a5a063326d2fb7cba