Fix bug saving trial via sqalchemy (#3690)
Summary:
Pull Request resolved: https://github.com/facebook/Ax/pull/3690

The issue here is that saving a trial was not idempotent. If you saved a trial with data and then saved the trial again, then the data would be deleted.

This seems really bad, but I'm honestly not sure of the scope of this bug. I discovered this while working on a test in the next diff in the stack, where a trial is saved with data and then the trial is saved again after being marked completed. The second save deleted the data from the first, and there was no Data on the experiment.

Using the ColumnOperator rather than python native `not in` resolves this. This was introduced in D54879641.

Reviewed By: esantorella, Balandat

Differential Revision: D73810845