Fix flaky sensitivity test
Summary:
Fixes a sensitivity test that was flaky because it relies on a
gradient estimate that is not deterministic for signing. There is already a
proper test for signing lower that accounts for nondeterminism in the gradient,
so we can have this test focus on what it is testing (second order measures)
and save the signing bit for the later test.

Differential Revision: D73920896