Fix flaky sensitivity test (#3706)
Summary:
Pull Request resolved: https://github.com/facebook/Ax/pull/3706

Fixes a sensitivity test that was flaky because it relies on a
gradient estimate that is not deterministic for signing. There is already a
proper test for signing lower that accounts for nondeterminism in the gradient,
so we can have this test focus on what it is testing (second order measures)
and save the signing bit for the later test.

Reviewed By: Balandat

Differential Revision: D73920896

fbshipit-source-id: a67577f7bdf7db7cffa22ead9ff76f9aa22bbbe1