Remove landing page link to quickstart tutorial (#3710)
Summary:
Pull Request resolved: https://github.com/facebook/Ax/pull/3710

The landing page is unversioned, so the `"/docs/tutorials/quickstart/"` link was actually pointing to stable but that tutorial is not yet in stable. This is breaking docusaurus's link checker blocking website builds.

The error is correct, we shouldn't be using links to "unreleased" pages from the landing page. Here we revert that link back to what's correct in stable.

I'll open a followup diff to change it back to the quickstart tutorial, that diff should be landed alongside the version release.

Reviewed By: mpolson64

Differential Revision: D73937615

fbshipit-source-id: c30366987a75885463c00f427483a1ef2105b007