Add CenterGenerationNode to choose_generation_strategy (#3708)
Summary:
Pull Request resolved: https://github.com/facebook/Ax/pull/3708

We have used center point in all benchmarks in the paper and decided to include it in Ax going forward. This diff adds it to the `choose_generation_strategy` that is used by Ax API. Center point is added by default, but the user has the option to opt out using `GenerationStrategyConfig`.

The transition criteria was updated to ensure the `initialization_budget` respects the user input (or the previous default) even when the center point is added. The transition criteria for `min_observed_initialization_trials` was updated to always count all trials on the experiment. There isn't a good way to include only trials from Sobol & Center and it seems sensible to transition when enough training data is collected, regardless of where it originates from.

Reviewed By: esantorella

Differential Revision: D73063657

fbshipit-source-id: 5c884361eb9960a078785dd95b1de9a264d2d697