Rename model_bridge inputs to adapter in model fit quality helpers (#3720)
Summary:
Pull Request resolved: https://github.com/facebook/Ax/pull/3720

Just a tiny slice of the argument & file renaming that we need to do to see adapter / generator refactor to completion.

NOTE: This diff renames arguments in a BC-breaking fashion. I think this is ok since these are relatively low level helpers and shouldn't be directly used in user workflows.

Reviewed By: Balandat

Differential Revision: D74073117

fbshipit-source-id: c14c8823ba1c7305fe5d4d7c38e07ca3716a59f5