Move non-user facing configs to new structs module (#3743)
Summary:
Pull Request resolved: https://github.com/facebook/Ax/pull/3743

Rename the configs users do not interact with to "structs" and remove them from the API.

These will provide value in the future by mirroring our thrift layer and by serving as a centralized place for validation. They are removed from the API to allow for more flexibility.

Will update docs and tutorials at the top of this stack, so expect tutorial tests to be broken.

Reviewed By: lena-kashtelyan

Differential Revision: D74328845

fbshipit-source-id: c566e63f5d58eba9cf46d3ddc20d4741be292c12