No enums in Ax API (#3740)
Summary:
Pull Request resolved: https://github.com/facebook/Ax/pull/3740

Remove Enums (ParameterType, ParameterScaling) in the API in favor of string literals. Now there is no need to import an extra class and the method signatures look less scary overall.

What we gain here in terms of ease of use we lose in runtime safety (literals are checked at typecheck time but relax into str at runtime) but overall this seems like a good tradeoff.

Will update docs and tutorials at the top of this stack, so expect tutorial tests to be broken.

Reviewed By: lena-kashtelyan

Differential Revision: D74266083

fbshipit-source-id: 515951eed116e6d3f5d87fdbe977152f4a5e6012