Add workflow to test website build (#3769)
Summary:
Before this there was nothing stopping us from merging website-breaking changes. This was happening often and wouldn't get caught until our daily cron failed to build the website. Some of these errors can be caught by just running the website locally, but others require building the prod version of the website which involves merging in (but not deleting) the `docusaurus-versions` branch. It's much simpler if we can automate this as a workflow at PR creation.

I've configured this to run on PRs only when files directly relevant to the website build are modified (to reduce overhead for unrelated PRs). For the main branch this will be more thorough and run every time a PR is merged.

Pull Request resolved: https://github.com/facebook/Ax/pull/3769

Test Plan:
I opened a separate PR to trigger this new workflow as a test:
- PR: https://github.com/facebook/Ax/pull/3770
  - Workflow: https://github.com/facebook/Ax/actions/runs/14938483763/job/41971282706?pr=3770

Reviewed By: saitcakmak

Differential Revision: D74507665

Pulled By: CristianLara

fbshipit-source-id: d3f4a021376a2f8d984795def1dac14106fbc72d