Create an analysis for reporting metric fetching errors (#3730)
Summary:
Pull Request resolved: https://github.com/facebook/Ax/pull/3730

This change is to introduce a new health check for metric fetching errors. This is used to generate an AnalysisCard with a dataframe containing records of the error.

This change also introduces a new private attribute `_metric_fetching_errors` to `Experiment`, which contains a list of metric fetching errors along with additional metadata (metric name, trial number)

Differential Revision: D68590971