Fix missing points in ArmEffectsPlot (#3777)
Summary:
Pull Request resolved: https://github.com/facebook/Ax/pull/3777

Added missing filter for nan values in go.Scatter construction. We were correctly removing elements without a mean in the x array, but forgot to do this in the y array. This lead to the arrays being different lengths and some points effectively being "cut off" and failing to render even though they appeared in card.df

Reviewed By: mgarrard

Differential Revision: D74668723

fbshipit-source-id: 6c722a8b9636307d347a0091dc0b1123a9136fb6