If filtering the trial to plot, should we also use this target trial (#3782)
Summary:
Pull Request resolved: https://github.com/facebook/Ax/pull/3782

When I was updating the advanced tutorials to use the new analyses, I ran into a bug where we were filtering to only trial 0, but the target trial for the experiment would be trial 5.

V2 of this diff:
So, after thinking about this more last night I think we will be passing in the trial_index in situations where we want to still use the target trial for model predictions if possible - and likely this will be possible because the experiment has all the data, and the predictions use the expeirment object. Also this is fixing a niche adhoc use case so the more we can contain it away from core ax logic the better.

V2 of this diff updates the logic of where to grab the status quo values from, and uses the filtering only there. Essentially we are adding some fallback cases that should prevent user errors

V1 of this diff implemented this idea:
This diff adds an argument to our target trial selection util which will filter the trials we consider for target trial to only those

Reviewed By: lena-kashtelyan

Differential Revision: D73584615

fbshipit-source-id: 2f54e5ce5f157017884cf129a1167228e7b3b13a