Include auxiliary experiment datasets in surrogate model fitting (#3774)
Summary:
Pull Request resolved: https://github.com/facebook/Ax/pull/3774

## What this diff enables

With `PE_EXPERIMENT` in the auxiliary experiment, it now extract the dataset from the aux experiment and fit a submodel on the dataset, which may be used later (e.g., to construct a learned objective in the `Acquisition` class).

In the meanwhile, `self._model` is still modeling the original set of outputs (so that everything else such as cross validation or candidate generation still works as usual), except that there is now an additional preference model fitted and accessible in `self._submodels`.

## Notable Changes
- Support parsing `RaningDataset` in `TorchAdapter` (and as a bonus point, we may deprecate [PairwiseAdapter](https://www.internalfb.com/code/fbsource/fbcode/ax/modelbridge/pairwise.py?lines=24-92) as all the logic there is upstreamed (will do that in a followup diff).
- `choose_model_class` takes in just a single dataset instead of a list of dataset as we may have different default models for different datasets (e.g., `SingleTaskGP` for most datasets but `PairwiseGP` for `RankingDataset`)
- `mll_class` is no longer a required field in `ModelConfig` dataclass. When missing, it'll be inferred altogether with the default botorch_model_class.
- Removed the `default_botorch_model_class`, and instead fill in missing value of `ModelConfig` according to dataset and search_space_digest as we may need different models (e.g., PairwiseGP vs. SingleTaskGP) for different datasets.
- Updated `use_model_list` to not rely on `default_botorch_model_class`, which is something we might have to infer when fitting individual models.

Meta: [original design doc](https://docs.google.com/document/d/1KENayOLCe7-kT85qlOk09AXUHhJXLpqKhsgyVVTwa2Q/edit?usp=sharing)

Reviewed By: Balandat

Differential Revision: D73286263