only warn on MetricFetchE if trial is complete
Summary: It is expected that a metric may not be fetchable while a given trial is running, especially early on in that trial's run. This diff moves the associated warning to only fire when a metric is still not fetchable after the trial has completed, and raises the loglevel to an error.

Reviewed By: mpolson64

Differential Revision: D74897258