Create an analysis for reporting metric fetching errors (#3730)
Summary:
Pull Request resolved: https://github.com/facebook/Ax/pull/3730

This change is to introduce a new health check for metric fetching errors. This is used to generate an AnalysisCard with a dataframe containing records of the error.

This change also introduces a new private attribute `_metric_fetching_errors` to `Experiment`, which contains a list of metric fetching errors along with additional metadata (metric name, trial number)

Reviewed By: mgarrard

Differential Revision: D68590971

fbshipit-source-id: a8ac09d2ff31b2fd6a00c5cf5bd8fd7fafd218dc