Only log error when attempting to attach Err for trial in COMPLETED state (#3803)
Summary:
Pull Request resolved: https://github.com/facebook/Ax/pull/3803

This error is logged very often for still-running trials. I understand why this is here, but also would like to avoid it being so common. A few potential options:
1. Log an exception only if the corresponding trial is complete (this diff), or some other filtering condition is met. Otherwise log at the info level.
2. Filter Errs upstream of `attach_fetch_results`, alerting at that . I believe the present alerting logic would never be called in this case, so this seems like an overly aggressive suppression.
3. Combine exceptions to at least reduce log spew.

I'm open to other suggestions here as well, but I believe we will benefit users and ourselves if we can reduce these logs to raise in a more narrow set of cases where it is relevant.

Reviewed By: mpolson64

Differential Revision: D75091238