Rename model input and attribute of Adapter to generator (#3809)
Summary:
Pull Request resolved: https://github.com/facebook/Ax/pull/3809

Renames the input & the attribute and updates the docstrings where "generator" term seemed more appropriate.

There are many other methods and attributes with `model` in their name. I'll tackle them in separate diffs to keep both diff authoring and review manageable. The term "model" was very overloaded, so renaming it to "generator" requires quite a bit of care.

Reviewed By: Balandat

Differential Revision: D75173850