Support accessing the current generation node in ESS
Summary:
This diff allows the early stopping strategies to access the current generation node and the underlying components when making early stopping decisions. This can be useful if we want to utilize the surrogate model for early stopping, as it will allow us to access it directly without having to re-fit.

NOTE: Opted to go for current node here rather than the full `GenerationStrategy`. There has been discussions around merging ESS into GS, and using GNode here avoids the need for a refactor to avoid cyclical dependencies if we decide to go with that refactor.

Differential Revision: D75097822