Do not let different transforms in Adapter._transform_data share a SearchSpace instance (#3830)
Summary:
Pull Request resolved: https://github.com/facebook/Ax/pull/3830

**Context**: In D74846187, we saw that some tests were taking a very long time.

This is in large part due to a nasty mutability bug. The`IntToFloat` transform uses the wrong search space, causing it to try a new randomized rounding that doesn't work 10,000 times. The tests were printing many warnings about this. The search space is wrong because different transforms in `Adapter._transform_data` sometimes (not always) share a search space, as in they use the same Python object, so mutating one can change the search space on another transform. There is also a potential issue where they share `Parameter`s but not a search space. In this case, the culprit was `UnitX` and `IntToFloat` sharing a `RangeParameter` which `UnitX` mutated, leaving `IntToFloat` with a parameter in the transformed space and its constraint in the untransformed space.

**This PR**: Clones the search space between transforms so that it is not mutated in place.

Reviewed By: Balandat

Differential Revision: D75569748

fbshipit-source-id: b5efe71dd861b61e2de5d0c2ad89a533d6bdadd1