Avoid concatenating empty DataFrames to stop using deprecated Pandas functionality (#3837)
Summary:
Context: Unit tests are producing many of this warning:

```
[...]/ax/Ax/ax/core/data.py:293: FutureWarning:

The behavior of DataFrame concatenation with empty or all-NA entries is deprecated. In a future version, this will no longer exclude empty or all-NA columns when determining the result dtypes. To retain the old behavior, exclude the relevant entries before the concat operation.
```

This stems from trying to concatenate empty DataFrames.

This PR: Only concatenates non-empty DataFrames in `BaseData.from_multiple`. If there are none, it will fall back to the behavior used for an empty list, which is to create a `BaseData` with no `df` input.


Test Plan: Warnings from ax/analysis/plotly/tests/test_sensitivity.py are no longer there.

Differential Revision: D75630996

Pulled By: esantorella