Clean up sensitivity analysis tests and add characterization tests (#3845)
Summary:
Pull Request resolved: https://github.com/facebook/Ax/pull/3845

Context:

Characterization tests (sometimes called "golden master" tests) check that output matches exact values so that we can be confident that a refactoring has not changed behavior. These show that the refactoring in a subsequent diff does not change behavior. I am not sure they will work on every machine, but so far they seem to be working in the GitHub CI and Sandcastle.

This PR:
* De-duplicates logic
* Sets a seed
* Adds checks for outputs. Tolerances are fairly loose since this is not deterministic even with the seeds set.

Reviewed By: Balandat

Differential Revision: D75711174

fbshipit-source-id: ac73a1784f6db9cd9b790e138f6ff90e822273c3