Adding auxiliary experiment methods to Experiment class
Summary:
Adding auxiliary experiment methods to Experiment class to add (as first experiment) and find auxiliary experiment.

Meta: Add to first is needed as we currently support one aux PE experiment and the first one is the one with highest priority and is the one that will be used. Find aux experiment is needed to support pref opt config (see the diff on top), which will allow us to identify which PE exp to use instead of relying on the order.

Reviewed By: mgarrard

Differential Revision: D75696082