Rename gen to gen_single_trial (#3844)
Summary:
Pull Request resolved: https://github.com/facebook/Ax/pull/3844

This diff updates the current `gen` method to be called `gen_single_trial` -- we will update gen_multiple_trials_with_multiple_models to be called gen in the next diff (initially tried to do them in the same diff and it was very confusing).

I decided to replace everything that currently calls gen with gen_single_trial instead of replace some of them with gen_multi as i think it kept things cleaner. This primarily effects test files.

Reviewed By: lena-kashtelyan, esantorella

Differential Revision: D75688420